pyproject.toml
[project]
name = "technical-manual-rag"
version = "0.1.0"
description = "RAG system for technical user manuals"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "llama-cloud>=2.0.0",
    "llama-index>=0.14.19",
    "llama-index-core>=0.14.19",
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-llms-google-genai>=0.9.1",
    "llama-index-llms-vllm>=0.7.0",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
